fix(module8-task1): Удалил моки
