--- project/src/mocks/index.ts
@@ -1,2 +1,0 @@
-export { offers } from './offers';
-export { cities } from './cities';
